# Conflicts:
#	src/main/java/PROJET/AL/BIBLIO/demo/proxy/EmpruntProxy.java
#	src/main/java/PROJET/AL/BIBLIO/demo/repository/EmpruntController.java
#	src/main/java/PROJET/AL/BIBLIO/demo/repository/UtilisateurRepository.java

+++ src/main/java/PROJET/AL/BIBLIO/demo/entity/Emprunt.java
@@ -14,8 +14,8 @@ public class Emprunt {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private int utilisateurId;
-    private int livreId;
+    private Long utilisateurId;
+    private Long livreId;
     private String dateEmprunt;
     private String dateRetour;
